Debian: Fixed banner error message
When logging, a message would show saying
"-sh: The: command not found"

This was due to a warning message that was being added
to /etc/bash_completion.d/openstack when the command
"openstack complete" is executed to fill the file.

Added a "-q" flag to suppress the warning message.

Test Plan:
PASS normal banner when logging
PASS file does not contain the warning message
PASS bootstrap and unlock

Closes-Bug: 1978493

Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>
Change-Id: Ie935ae2c32e644cd6bc622046319adfb90d6477e