Add a puppet class to support k8s feature-gate update
This change adds a class to the kubernetes puppet manifest.
It executes a script that updates kubernetes kubeadm-config
configmap and static manifests and restarts the
kube-api-server and kube-controller-manager with new
feature gates. The class ensures that the api-server and
the controller-manager pods are back up and healthy.

Depends-On: https://review.opendev.org/c/starlingx/integ/+/845654

Test Plan: (CentOS)
On AIO-SX and AIO-DX:
PASS: Full platform upgrade successful
PASS: Verify platform logs and puppet logs for
      successful execution.

Story: 2009789
Task: 45626

Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>
Change-Id: Ia789a84bf012b3e574e9319eb68293857628db28