Fix umask not found when applying kubernetes.pp
On Debian, umask is a built-in command so "umask 077" in
kube-apiserver-change-params.erb will fail the manifest
since umask could not be found. This change fixed it by
running umask with shell as provider.

Test Plan on Debian and CentOS:
PASS: "system service-parameter-apply kubernetes" complete
      successfully, no error in puppet.log.
PASS: check /tmp/cluster_configmap.yaml and
      /tmp/kubeadm_config_view.yaml have the right permission:
      -rw------- 1 root root /tmp/cluster_configmap.yaml
      -rw------- 1 root root /tmp/kubeadm_config_view.yaml

Closes-Bug: 1978718
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: Ie680cac904ef10bf063c0e6329192edc075e6a1e