Modify kubernetes_local_secrets_audit interval
This audit only purpose is to ensure synchronization between k8s
secrets and keyring, but there's already a notification system by
keystone that sync this data.

After a research, we noticed that the appropriate time interval
for this audit is to be executed only once per day.

This commit is a continuation of the study started at task 45711.

PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45720

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/848329
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I334c2900ba5a824e80988516f179e9d92ceb87bb