Fix stale data cleanup on the logical volumes
The stx-puppet tries to wipe the start and end of some logical
volumes to prevent problems, but the command uses the wrong output
file resulting in the creation of junk files. The command for scratch
does not work since it's already mounted before the bootstrap. For this
reason it was decided to unmount the filesystem before wiping the
device's data.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/849047

Test Plan:
PASS: Perform bootstrap and restore-platform playbooks on centOS
PASS: Perform bootstrap and restore-platform playbooks on Debian

Closes-bug: 1975851

Signed-off-by: Hugo Brito <hugo.brito@windriver.com>
Change-Id: I3d0aa664a1bd72dd49ca2f4941bc06d25517b474