Modify _audit_device_image_update interval
This method verifies the timeout of the update process of the images,
and the minimum timeout available inside the code is defined in the
fw_update_small_timeout variable. This variable is not changed by any
other process and its value is set to 300 seconds.

After a research, we noticed that the appropriate time interval
for this audit is to be executed between 300 seconds interval.

This commit is a continuation of the study started at task 45711.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45777

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/848329
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: I50d19f9b6b930b2d6b587eb811cd6e6ed33bce2f