Modify _storage_backend_failure_audit interval
This audit verifies if there's a storage backend stuck in 'configuring'
state, the timeout for this is 1200 seconds, set in the
SB_CONFIGURATION_TIMEOUT constant.

After a research, we noticed that the time interval does not need to be
60 seconds, since the timeout of the configuration is 1200. Increasing
it to 400 seconds (1/3 of the timeout) is enough to capture any storage
backend stucked.

This commit is a continuation of the study started at task 45711.

TEST PLAN:
PASS: AIO-SX: manually replaced these files into a Debian installation
and no crashes happened.
PASS: rebuild the whole system with the modification with no crashes.
PASS: AIO-SX: verify that the audit is being called during the period.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: guarantee that all audits are being called by the periodic tasks.

Story: 2010087
Task: 45828

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/848329
Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: Idafdf98e0dfd6942b31284a3fd1b47c649c82634