Modify the audit interval of upgrade status to 180 seconds
The audit_upgrade_status method verifies if there is a system upgrade
running. If so, it takes actions based on its state. Note that the
system stays in each state for a long time but the periodicity of this
audit is 60 seconds, causing unecessary executions during each state.

After a research, we noticed that the appropriate time interval
for this audit is to be executed between 180 seconds interval. This
interval was choosen because upgrades are not frequenty done on the
system and they already take time to be executed.

This commit is a continuation of the study started at task 45616.

TEST PLAN:
PASS: rebuild the whole system into a fresh new ISO. Install,
bootstrap, and unlock it into a system with no crashes.
PASS: AIO-SX: verify that this audit and also all other audits are
being called during their periods. This process was done analyzing the
sysinv debug logs inside the system. From it, its possible to verify
all the audits being called.
PASS: follow the sysinv logs seeking for errors. No error was found.
PASS: AIO-SX: perform a system upgrade with this new interval.
PASS: AIO-DX: failed to perform a system upgrade with this new interval
(was expected to fail).

Story: 2010087
Task: 46000

Signed-off-by: Bruno Costa <bruno.costa@windriver.com>
Change-Id: Ia9b1b83face84e60c879fffb420e51a666a7c40c