Debian: apply network config with correct dependency
It was detected that on a system with 2 or more labeled interfaces
(those are ifupdown representation of "interfaces" that will only
configure the IP address on a base interface, e.g. interface vlan160
may have labeled interfaces vlan160:1 and vlan160:5) the order of
execution was preventing them to be applied because the execution
was happening before the base interface.

The execution was decided by the result of the find command, this
change applies a sort on the list to be sure that the execution is
respecting the base interface first, e.g:
i)  enp0s8 => enp0s8:1 => enp0s8.100 => enp0s8.100:1 => enp0s8.100:5
ii) vlan160 => vlan160:1 => vlan160:5

Also:
1) added a log message in case of ifup or ifdown failure
2) prevent the processing of labeled interfaces in
   verify_all_vlans_created()

Test Plan (Debian and CentOS):
PASS add interfaces and lock/unlock the controller
PASS edit interfaces and lock/unlock the controller
PASS remove interfaces and lock/unlock the controller

Closes-Bug: 1987918

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I7bc8d2bb25a74a5900b46936b59fc83b9251b53b