Debian: labeled interface stanza creates deprecated IPv6 address
Labeled interface stanzas for inet6 family are creating deprecated
addresses, due to ifup internal logic (it is setting preferred_lft to
zero). If an IPv6 address is marked deprecated it will not be used as
source address for communications from the interface.

This change makes the parser identify these cases and sets the label
interface to the base interface for the first label (e.g. :1) used by
sysinv to identify as platform networks when generates the
interfaces file in /var/run/network-script/

Test Plan (Debian only)
[PASS] install and unlock AIO-DX with IPv6 network, using management
       and cluster-host on the same interface vlan.
[PASS] executed a follow up lock/unlock to use the generated scripts
       in /etc/network/interfaces.d/
[PASS] add/edit/remove interfaces and then lock/unlock

Story: 2009968
Task: 46251

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I9ee27bb9181d196559c3e0143c3e4d859ca0413a