Fix error add parameter with invalid arg in K8s config
These changes fix the use case "add a valid
parameter with invalid argument" for kubernetes
custom configuration support. Currently, after
apply the changes, the Kubernetes cluster
crashes and won't recover.

It is necessary to restart kubelet because when an wrong
configuration is set in any of the k8s control plane
components, kubelet makes a maximum of 5 attempts to
restart the erroneous component. If we reach the limit,
no matter if we correct the configuration during the
automatic recovery, the container does not starts again.

The puppet timeout is increased to contemplate the
time to update the different k8s control-plane components
and the automatic recovery process, if necessary.

Test Plan:
* CENTOS and DEBIAN distro:
  - Fresh Install with AIO-SX and DX/STD.
  - Add new valid parameter with invalid arguments.
  - Apply changes on kubernetes service.
  - Verify cluster health and configuration.

Closes-Bug: 1992207

Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>
Change-Id: Ieb17f9ff7359813f60b8baffb3d8f53fe1e2d7f8