Fix installation of default ipv6 route
For a worker node in a Standard IPv6 system, a default route is
added to the system controller.

Currently, this default route will appear in the
/var/run/network-scripts.puppet/routes6 file, with it's contents
merged with the /var/run/network-scripts.puppet/routes file.

Ultimately, the combined routes will be written to Debian's
/etc/network/routes file.

An issue was noted in which the default route was not present
on a worker node after installation.  This led to the puppet
manifest failing when it tried to access the k8s registry on
the controller.

It was noticed that in the route merging code, an individual
route line was being appended to the end of the 'routes' file
with 'cat' rather than something more appropriate like 'echo'.

Closes-Bug: 1995142

Testing:
PASS: Ensure the default route is added and applied by the
      apply_network_config script.
PASS: Ensure previous failing puppet manifest now passes
PASS: Ensure the worker node can be locked/unlocked successfully
      and the default route retained.

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Ideaea85dbae05fea648f1cff5a12c2ddadf82f87