[PTP] Remove subsystem identifier check for PTP sysfs commands
The subsystem identifier for Intel E810 NICs is not a suitable
indentifier for determining whether a given NIC will support the sysfs
command that is trying to be set. This identifier can vary by adapter
type and manufacturer, with a wide range of values being possible. This
has lead to the configuration failing to apply on NIC variants that do
not have matching subsystem IDs, with no error being raised.

Rather than checking the ID before applying the sysfs command, simply
apply the command and return the error if it does not work. This allows
the configuration to be applied on all E810 NICs regardless of subsystem
ID and results in an error if there is an attempt to apply to an
unsupported NIC type. This has the benefit of alerting the user that the
supplied config is incorrect via the config-applied failure alarm and
allowing it to be corrected.

The error cases are:

Invalid arguments for an E810 NIC
-bash: echo: write error: Invalid argument

Attempt to apply command on a non-E810 NIC
-bash: /sys/class/net/enp24s9f2/device/ptp/ptp0/pins/U.FL1: No such file
or directory

Test plan:
PASS: Apply manifest to E810 NICs with various subsystem ids.
PASS: Manifest fails to apply configuration to unsupported NIC, alarm
raised, no other system impact

Closes-bug: 1995308

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I4d9e855a79ca54df1e08b55adcd8aaf2a5096e9b