Fix IPv6 corrupted address after system unlock
Currently, for an IPv6 system using multi-netting (for example,
mgmt, cluster-host sharing an interface), multiple aliased
interface files are created under
/etc/network/ifcfg-<ifname>:<label>, with the label being the
network-id.

Currently, we remove the label from the interface stanza in
these files in order to prevent the primary address (label :1)
from being marked as 'depreciated' by the Linux kernel.

It was noticed that the current label removal logic is too
broad.  It would also 'corrupt' an IPv6 address that had the
substring ':1'.  For example:

fdff:719a:bf60:1021::3 -> fdff:719a:bf60021::3

The invalid address was of course not accepted by the kernel,
and caused a communication issue over whichever network the
address belonged to.

This issue was seen after first unlock in a subcloud in a
distributed cloud environment.

The fix makes the 'sed' replacement more surgical.  We already
track the 'name' of the interface with both the label present
(iface_name) as well as without the label (base_iface_name)

This affects Debian only.

Closes-Bug: 1995395

Testing:
- On an affected subcloud, apply the 'sed' fix after bootstrap, but
  before first unlock.
- Ensure the IPv6 address in /etc/network/interfaces.d/<file> is
  correct (match with bootstrap params and DB) after unlock.
- Ensure the subcloud becomes 'online'.

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Ic3461295af41dd6a335c103c78522ca8371f73ec