Fix kubernetes options during upgrade
The kubernetes upgrade of control-plane was failing due to
incompatible settings in kubelet_override.yaml. Unless all
overridden options are specified (some can be k8s version specific),
the 'kubeadm upgrade apply' reverts back to the current defaults,
and subsequent 'kubeadm upgrade node' may fail.

This issue manifested when --config kubelet_override.yaml was first
introduced to deal with 'cgroupDriver: systemd' becoming the upstream
default. We provided this mechanism to override the settings
(i.e. to cgroupfs instead of systemd).

The kubelet_override.yaml is currently empty and not required.
If we do not specify an override file, the current kubelet config
settings are preserved over the upgrade, which is what we want.

The --kubeconfig option is removed since admin.conf doesn't work
with 'kubeadm upgrade node' and default is sufficient.

Story: 2010301
Task: 46724

Test Plan:
PASS: Standard: Kubernetes upgrade 1.23.1. to 1.24.4
PASS: AIO-SX: Kubernetes upgrade 1.23.1. to 1.24.4

Signed-off-by: Jim Gauld <james.gauld@windriver.com>
Change-Id: I0d61f2aee07ec63d18279a3e80fe2663c3dbd48b