Add vgchange before wipe new PV
When we have a nova-local on a separate disk, after a fresh
install (without wiping the disk), it's not possible to wipe
the disk because that disk is "busy". So the vgchange command
was added before the wipefs to make sure that the volume group
is not busy.

Test Plan:
PASS: Fresh install AIO-SX with a separate disk for nova-local and
repeat fresh install without wiping the disk.

Closes-Bug: 1995506

Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>
Change-Id: If3da66734dc1121d3d2a27de804c0194a6d5e0fe