robustify AIO-SX kubelet upgrade
During a kubelet upgrade, we need to change the stage2 bindmount
from the old version of kubelet to the new version.  Unfortunately
we have seen this intermittently fail due to the stage2 mount
being "busy" due to kubectl running at the time of the attempted
unmount.

The real solution will likely be to change to using symlinks
instead, but for now we can improve robustness by explicitly
unmounting the stage2 filesystem with retries, then explicitly
modify the bind mount, then explicitly mount the stage2 filesystem
again.

We also change the order of operations slightly to mask the kubelet
service before shutting it down in pmon.  This avoids any chance
of the service being restarted spuriously.

We also add the "--runtime" flag to the unmount/mount operations.
That way if anything goes wrong a reboot can restore the stage2
mount.

Test-Plan:
PASS - Perform a normal manual K8s upgrade to 1.24
PASS - Repeatedly trigger a standalone version of the puppet
       manifest to alternate between the 1.23 bind mount and the
       1.24 bound mount.  Run for several hours without failures.

Story: 2010301
Task: 46896
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Iffc4a09cb0b15022aa541d85e06624016efd5374