Restart sssd service on ssl_ca install
The installation of WAD CAs is not taking effect because the sssd
service is not getting restarted after the certificate is installed.

This commit adds the restart of the sssd service to
platform::config::certs::ssl_ca, which is the puppet class that gets
executed when a new ssl_ca certificate gets installed.

Test plan:

PASS: Add a ssl_ca certificate with system certificate-install and
      verify that the sssd service gets restarted in sequence.
PASS: Uninstall a ssl_ca certificate with system certificate-uninstall
      and verify that the sssd service gets restarted in sequence.
PASS: In a DC environment verify that ssl_ca updates from the system
      controller get pushed to the subcloud and that the sssd service
      will then get restarted in the subcloud as well.

Closes-Bug: 1997881

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: If1eace8849f752022bd58af47de12771e893590f