Adjust the min_free_kbytes setting way for storage
We already have the approach to set the sysvar vm_min_free_kbytes value,
reserving 128MB for every 25GB of system memeory, with also a minimum
amount specified that we won't go below.
But we have a customer with a 64GB storage node and their tuning
experimentation shows that they need 1GB reserved in order to avoid OOM.
We increase the value for storage node and adjust the method to reserve
256MB for every 16GB of system memory, It tries to keep just
approximately 1.5% of memory reserved. So this should get the customer
set at 1GB if they have 64G of total RAM.

TestPlan:
Pass: build-pkgs
Pass: build-iso
Pass: jenkins installation
Considering it's difficult to find a storage node to verify if the value
of min_free_kbytes is expected, I move the same puppet syntax change for
storage in sysctl.pp to the controller area, then check the value of
controller node as follows:

controller-0:~$ free -g
              total   used   free   shared   buff/cache   available
Mem:            125     84     35        0            4          35
Swap:             0      0      0
controller-0:~$ cat /proc/sys/vm/min_free_kbytes
1835008

We get the min_free_kbytes value as expected.

Closes-Bug: 1940855

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: Ieaa6ea93537ed066c8b764d4b6e5e1ff3965f310