Add '/etc/pki' item for puppet ensure directory
The ssl_ca installation is failing for worker nodes because it does not
find the /etc/pki folder created already.

This commit adds /etc/pki individually as an item to a puppet ensure
directory statement. This will make sure that puppet creates this
directory first when it does not exists. If the dir already exists
puppet will also be satisfied and execute with success.

Test plan:

PASS: Add a ssl_ca certificate with system certificate-install and
      verify that certificates where added to
      etc/pki/ca-trust/source/anchors/ca-cert.crt in a compute node.
PASS: Run a full deploy of a standard lab and verify that compute nodes
      become unlocked and available.

Closes-Bug: 1999588

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: Ib59ab88a9d4d1112e35f98d92aef72cbac01af07