Make '/etc/pki' and children directories 755
It was suggested in a code review that directories should be 755.
To keep consistency with previous stx versions as well, it makes sense
to have change it to 755 in order to avoid unintended side effects.

This change affects the whole dir hierarchy:

/etc/pki
/etc/pki/ca-trust
/etc/pki/ca-trust/source
/etc/pki/ca-trust/source/anchors

Test plan:

PASS: Add a ssl_ca certificate with system certificate-install and
      verify that certificates where added to
      etc/pki/ca-trust/source/anchors/ca-cert.crt in a compute node.
PASS: Run a full deploy of a standard lab and verify that compute nodes
      become unlocked and available.

Closes-Bug: 1999664

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I2a8bdbe1ca4b13275f85ef54bf261213cc6728bf