Add SM support for the DC admin network
This commit adds SM support for the DC admin network.

The admin network is intended to be used between a subcloud
and system controller. Because the (existing) management network
is so embedded in other parts of the StarlingX system, it makes
it prohibitively hard to re-configure this network after initial
installation.  The admin network is intended to be isolated from
the management network, allowing re-configuration of the network
parameters in the case that the physical network between subcloud
and system controller has been changed.

In the case of admin network usage, the management network still
exists but is a private network in the context of a subcloud.

This commit provides for SM parameters of the admin network
to be updated at runtime for an AIO-SX subcloud.

Depends-On: https://review.opendev.org/c/starlingx/ha/+/868361

Testing:

  AIO-SX:
    - admin-ip service is enabled when the admin network is
      created.
    - admin-ip service is not enabled when the admin network is
      not created.
    - floating-ip is updated on the admin interface when admin
      addr-pool information is changed.
  AIO-DX:
    - admin-ip service is enabled when the admin network is
      created.
    - admin-ip service is not enabled when the admin network
      is not credated.
    - floating-ip is updated on the active-controller when the
      admin addr-pool information is changed.
    - When a peer admin interface is down, an alarm is raised.
    - When a peer admin IP is not correct (changed), an alarm
      is raised.
    - Swact between controllers.

Story: 2010319
Task: 47277

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Idd22af3c9c8c95586407bd0a6394d944aeca55a8