Update k8s plugin directory
We need to support upgrading from a CentOS-based system
to a Debian-based system.  Debian uses OSTree which
means that /usr is read-only.

During the upgrade of a multi-node system there is
an issue where the kube-controller-manager-controller-1
pod (created on the CentOS system) is trying to write to
/usr/libexec which is not valid on the upgraded node running Debian.

This commit is to update the pod's yaml file with the new
writable directory /var/opt/libexec after Debian is installed.
The pod is retarted after the yaml file is updated.

A separate change will be made to update kubeadm-config
configmap with this new directory on Debian system. This
configmap change will make sure the new directory is
used permanently.

This commit only covers platform upgrade case. Downgrade is not
covered and there will be a separate commit for it.

Test Plan:

PASS: Upgrade from stx 6.0 to stx 8.0 in AIO-DX and the pod is
running
PASS: Upgrade from stx 6.0 to stx 8.0 in AIO-SX and the pod is
running
PASS: Fresh Debian iso installation on DX

Task: 47161
Story: 2009303

Signed-off-by: Junfeng (Shawn) Li <junfeng.li@windriver.com>
Change-Id: I503beef44a7c1a16a12226a2988e54535c5c4162