Deactivate provision of the guest-agent service
The guest-agent service it is currently being activated
in setups where stx-openstack is applied but it's not
being used since we went to containerized openstack.
Since this service is no longer being used and it is
currently causing coredump issues [1] in debian environment
we are deactivating the provision of the service.
The service can be reactivated in the future if needed (we
created a StoryBoard [2] to keep track of this modification).

[1] https://bugs.launchpad.net/starlingx/+bug/2000168
[2] https://storyboard.openstack.org/#!/story/2010520

Test Plan:
PASS: Generate the debian image without the code that
performes the provision of the guest-agent service.
PASS: Verify that after applying the modification in a host
WITH stx-openstack applied the guest-agent service it is not
being provisioned.

Depends-on: https://review.opendev.org/c/starlingx/tools/+/870433
Depends-on: https://review.opendev.org/c/starlingx/nfv/+/869817

Partial-Bug: 2000168

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I914ad7caa241bcd42ae1393d943b533a4ad32c06