Fix pod rollout script of K8s root CA change
The fixes and improvements of the pod rollout script used in Kubernetes
Root CA update procedure are listed below. This script is used in the
stages "trustbothcas" and "trustnewca" of the procedure. It is written
in the file "kube-rootca-update-pods.erb".

In file "kube-rootca-update-pods.erb":

Fix: the variable "statefullsets" had a typo and was changed to
"statefulsets". Because of this typo, in line 29, this variable was
being used without being properly initialized. When the system had at
least one statefulset, the script always got stuck trying to restart a
statefulset inside the non-existing namespace "0", the internal script
timeout was always reached and no existing statefulset was restarted.

Improvement: the internal script timeout was increased from 8 minutes
(32 retries every 15s) to 30 minutes (40 retries every 45s).

Fix: It was noticed that "kubectl rollout status --watch=false" output
does not always contain "successfully" in it when the rollout is
completed, like the example below:
"partitioned roll out complete: 2 new pods have been updated...".
It also always output "0" as command result. Because of this, it was
used "--timeout=100ms" instead of "--watch=false", to make the command
output "1" when the rollout was not completed and "0" when it was.

Improvement: more logs were added to the script.

Fix: The script now returns error when the internal script timeout is
reached and the rollout was not completed. Before, it simply exited and
the procedure would continue not pointing problems.

In file "kubernetes.pp":

Improvement: The hard timeout of Puppet for the pod rollout script was
changed from 10 minutes to 60 minutes. This timeout should not be
reached, as the internal script timeout is around 30 minutes, however it
is not impossible, as the 30 minutes timeout doesn't count the time
spend in the execution of "kubectl rollout status" commands.

Test Plan:

PASS: In a AIO-SX with at least one statefulset, being these
statefulsets with no problems, execute the Kubernetes Root CA update
procedure and check that it is completed successfully without reaching
any timeout in stages "trustbothcas" and "trustnewca".

PASS: In a AIO-SX with at least one statefulset, being these
statefulsets in error state, execute the Kubernetes Root CA update
procedure and check that it transitioned to an error state due to the
pod rollout script reaching the internal script timeout (that is around
30 minutes) and returning error.

Closes-Bug: 2004594
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I0c3fb7943dcecbd206cc09dbc122c2a0332213ba