Mask credentials in puppet log files
This commit masks the credentials used in crictl pull command, logged
and executed by puppet to pull images from registry. Wrapping the data
type of local_registry_auth variable as Sensitive, protects from
exposing the registry credentials in log files.

Test Plan:
PASS: Deploy a DX system with the wrong credentials addressed to the
      $local_registry_auth and observe in the puppet logs from
      Controller-1, the exec commands are logged with Sensitive
      [redacted] where it would have the credentials to the registry.
PASS: Deploy a DX system with the correct credentials addressed to the
      $local_registry_auth and observe the system successfully being
      available.

Closes-Bug: 2008726
Signed-off-by: Manoel Benedito Neto <manoel.beneditoneto@windriver.com>
Change-Id: I831e65ad948e62c346f56f5e9c7587e4dbdcd29f