Pull images from registry with credential file
This commit adds the ability to use the content of a created credential
file as the value of the --creds parameter for `crictl pull` command,
preventing logging the literal credential values on puppet log files.

The registry_credentials file is created from a template file that
interpolates the scope variable containing the value of the
credentials. As soon as the file is created and read, the file is
deleted from the /tmp/puppet directory.

Test Plan:
PASS: Deploy a SX system and observe Controller-0 status is available
      enabled active.
PASS: Deploy a DX system and observe in the puppet logs from
      Controller-1, the exec commands are logged with the sed command
      where it would have the credentials to the registry. Observe the
      both controllers are available enabled active status.
PASS: Observe exit 0 output in each log file of Controller-1 listed by
      the execution of the command below.
      find /var/log/puppet -type f | xargs grep -l "creds"
PASS: Upgrade kubernetes version from v1.24.4 to v1.25.3 in AIO-SX and
      AIO-DX systems. Observe the state upgrade-complete after finishes
      the manual kubernetes components upgrade for both types of
      systems.

Closes-Bug: 2008726
Signed-off-by: Manoel Benedito Neto <manoel.beneditoneto@windriver.com>
Change-Id: Id2579767f25cb5d0cabe38528e7aa72bb9c7a8b2