Add code that erases routes file prior to its content generation.
When generating new contents for the file '/var/run/
network-scripts.puppet/routes', Puppet reads routes that already
exist in the file and adds them to the new contents. This behavior
causes deleted routes not to be erased from the OS. This commit adds
code that erases routes from the file prior to it's content
generation, ensuring that its route list matches sysinv's.

Test plan

For each test, the following steps were performed:
1. Run route related command.
2. Check that changes were applied to the OS.
3. Lock / unlock host.
4. Check that changes remain active in the OS.

PASS Add route A. Check that route A is present in the OS.
PASS Add route B. Check that routes A and B are present in the OS.
PASS Delete route A. Check that route B is present in the OS and
     route A is not.
PASS Delete route B. Check that routes A and B are not present in
     the OS.

The tests were performed in the following systems:
  AIO-SX IPv4
  AIO-SX IPv6

Closes-Bug: #2015102

Signed-off-by: Lucas Ratusznei Fonseca
               <lucas.ratuszneifonseca@windriver.com>

Change-Id: Iec4bc1521c42dfd87da4085d80ea03f08028c950