Ignore platform_nfs cmds if plat_nfs_ip is empty
During an upgrade from Rel. 7 to Rel. 8 the upgrade-activate failed
and after some time it was executed again. In the second
upgrade-activate execution, the update_platform_nfs_ip_references
was called again, but it failed because plat_nfs_ip was empty.
It happened because this function was already executed once and at
the end of it the plat_nfs_ip is removed from system.yaml
So if this variable is empty, it means that upgrade-activate was
called again and there is nothing to do.

Test plan
PASS AIO-DX upgrade from Rel 7 to 8
PASS manual change in this function changing the plat_nfs_ip
     value to:
     $plat_nfs_ip = undef
     $plat_nfs_ip = ''
     And executing the upgrade-activate and confirm that
     commands were not executed

Partial-Bug: #2015392

Signed-off-by: Fabiano Mercer <fabiano.correamercer@windriver.com>
Change-Id: I1e9a9e31ab33b1d88125daf1195e8ac14b104e17