Remove references to deprecated script
upgrade_k8s_config.sh has been deprecated and
removed due to lack of support for "flow" style YAML.
Deprecated functionality has been superseded
by better YAML-aware handling in sysinv.

Updating how we invoke kubeadm, we will now use an explicit
version of kubeadm when calling it.  The version called
will now match the version we are upgrading to in order to handle
the format unsupported by previous versions of kubeadm.

Test Plan
PASS:
- Manually update scripts on controllers and worker nodes based on
  https://review.opendev.org/c/starlingx/integ/+/880390
- Perform manual upgrade from k8s v1.21.8 to v1.22.5
- Verify kubernetes successfully upgraded to v1.22.5

Test was performed in the lab with local changes
to verify the code.
Patch was not tested.

Closes-Bug: 2017696
Change-Id: I840eb566057be495fe0da3cae7604bf8055c0d4f
Signed-off-by: Gleb Aronsky <gleb.aronsky@windriver.com>