Ensure parameters are set for vfio-pci
If a pci-sriov VF interface is configured prior to configuring a host
device to use vfio-pci PF driver, the reload of vfio-pci module with
extra required parameter does not set the parameters because the module
is already loaded.

The fix is to set the parameters (enable_sriov and disable_idle_d3) by
writing to the parameters files in sysfs after the module is loaded.

Test plan:
PASS: With a pci-sriov VF interface provisioned, set a host device to
      use vfio-pci PF driver. Ensure no puppet errors and unlock
      successfully.

Closes-Bug: 2016698

Signed-off-by: Teresa Ho <teresa.ho@windriver.com>
Change-Id: I07425a333916fe5d05316baa33e82dd3d5b79699