Add L3 firewall support for platform networks
Adding puppet classes to install L3 firewall in cluster nodes that
can run kubernetes (controllers and workers), It uses the hash2yaml
function from the package puppet-hash2stuff, the change is marked as
a dependency for this task.

The story 2010591 is still under development and for now we are only
applying the platform firewalls into the controller nodes.

With the change https://review.opendev.org/c/starlingx/config/+/881495
the new classes' config info is provided. At this first delivery the
firewall will not contain restrictive rules, focusing more in making
the necessary GlobalNetworkPolicy and HostEndpoints to be correctly
installed among the nodes

Test Plan:
[PASS] install AIO-DX
[PASS] install Standard with DX+worker+storage nodes

Story: 2010591
Task: 47954

Depends-On: https://review.opendev.org/c/starlingx/integ/+/881497
Change-Id: I1d35abde612cdaf3ccb54a858618037382ff2636
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>