remove reference to deleted exec resource
In commit 77e0c7c1 we removed an exec resource that called out to an
obsolete script.  However, we neglected to remove a "require"
metaparameter which referenced the removed script.  It's unclear how
this was missed since the previous change was tested in VirtualBox.

This causes a puppet error when trying to upgrade K8s:

Could not find resource 'Exec[update kubeadm-config]' in parameter
'require' (file:
/usr/share/puppet/modules/platform/manifests/kubernetes.pp, line: 813)

The fix is to remove the metaparameter.

TEST PLAN:
PASS: While running the dev branch on AIO-DX, upgrade K8s from 1.21
      to 1.22.
      (note, a workaround was required to deal with
       https://bugs.launchpad.net/starlingx/+bug/2018247)

Partial-Bug: 2017696
Change-Id: I66c0e88f0f0a3acc3326391263123e60667561cc
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>