Fix k8s custom configuration configmap init flag
After system bootstrap, when service-parameter-apply is executed
for the first time, it verifies that the k8s configmaps linked to
service-parameters extra-volumes exist, then creates a flag
(configmap_initialization_flag) to skip this step on subsequent
runs. If the flag is not generated, the k8s custom configuration
script checks the k8s configmaps each time it is run.

Test Plan:
  PASS: Fresh Install STD/DX
  PASS: Apply K8s service-parameter.
  PASS: Verify configmap initialization flag has been created.

Closes-bug: 2022983

Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>
Change-Id: Ie28247fd62945f90a9018a7ebb7942245ea5aeb4