Create the OAM firewall with the new implementation.
This change moves the OAM firewall to be configured with the same
solution used for the other platform networks. The necessary yaml
config data is generated by sysinv.puppet.platform_firewall.

During tests, it was observed that the bring-up manifest did not
update the firewall config as it requires kubectl API to be available
at the moment of execution. When this happens the file
.platform_firewall_config_required is created in /etc/platform/
to trigger sysinv to apply the configuration using
platform::firewall::runtime.

Test Plan:
[PASS] Modify configured firewalls with kubectl edit and execute
        lock/unlock, after bring-up, check they were reset to the
        original config
[PASS] Erase a HostEndpoint and execute lock/unlock, bring-up check
        the HostEndpoint was reset to the original config.
[PASS] Erase a GlobalNetworkPolicy and execute lock/unlock, bring-up
        check the GlobalNetworkPolicy was reset to the original config.

Story: 2010591
Task: 48278

Change-Id: I6ff22310fe1c49c9a3e53d616b98879ee5469b6e
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>