Avoid unintended ldap configuration in subcloud
LDAP should not be configured in a subcloud as it connects to the
systemController's openldap service. The ldapserver_remote variable
is used to determine whether to configure a local ldap service or
connect to a remote one (to the system controller in this case).

Test plan:

PASS: Full build, install and bootstrap on a DC system and
      one subcloud added.
      Verify that the ldap directory is not configured in the subcloud,
      meaning that cn=config.ldif is missing in /etc/ldap/schema.
      Verify that slapd.conf is still present in /etc/ldap
PASS: System host-reinstall in a standalone DX environment.
      Verify local ldap is properly reconfigured at unlock after
      host-reinstall.
      Verify that slapd.conf is no longer present in /etc/ldap
PASS: System host-reinstall in a DX subcloud environment.
      Verify host-install works and local ldap server is not configured.
      Verify that slapd.conf is still present in /etc/ldap.
PASS: Create a ldap user in system controller and verify it's possible
      to login to the subcloud using that user.

Partial-Bug: 2025663

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I85de8e2ced2490d07823e285608d6e9b1e5e472a