Populate DNSMasq with host entries
Host DNS records must use the ".internal" domain
to configure FQDN entries with the format of <host>.internal.

Given the complexities of supporting the multi-host reconfiguration,
this task will focus on support for a simplex system only.

This commit focuses on the addition of cname records
related to the management config

Work Items:
- addition of cname records to dnsmasq.conf for AIO-SX

Testing:
- build successful
- succesfully bring up AIO-SX and AIO-DX
- ensure changes only apply on AIO-SX
- ping system host's new aliases (with and without ".internal")
  -> also with manual removal of mgmt entries from /etc/hosts
- ensure all mgmt interfaces are up
- ping registry.local

Story: 2010722
Task: 48295

Change-Id: Ie6334f4cf533699a93c406428696921e2e4d85bc
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>